fix folders naming
