+++ project/bot/src/main/java/ru/tinkoff/edu/java/bot/controller/RestResponseEntityExceptionHandler.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.scrapper.controller.exceptionhandler;
+package ru.tinkoff.edu.java.bot.controller;
 
 
 import org.openapitools.model.ApiErrorResponse;
@@ -8,16 +8,14 @@ import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import ru.tinkoff.edu.java.scrapper.exception.IncorrectRequestParameterException;
-import ru.tinkoff.edu.java.scrapper.exception.ResourceNotFoundException;
+import ru.tinkoff.edu.java.bot.exception.IncorrectRequestParameterException;
+
 
 @RestControllerAdvice
-public class RestExceptionHandler extends
+public class RestResponseEntityExceptionHandler extends
         ResponseEntityExceptionHandler {
 
     @Override
@@ -26,24 +24,12 @@ public class RestExceptionHandler extends
                                                                   HttpStatusCode status,
                                                                   WebRequest request) {
         ApiErrorResponse response = new ApiErrorResponse();
-        response.setDescription("Incorrect request body");
-        response.setCode(String.valueOf(HttpStatus.BAD_REQUEST.value()));
+        response.setDescription("Incorrect JSON");
+        response.setCode(String.valueOf(HttpStatus.BAD_REQUEST));
         response.setExceptionName(ex.getClass().getName());
         response.exceptionMessage(ex.getMessage());
         response.setStacktrace(ex.getStackTrace());
         return new ResponseEntity<>(response, status);
-    }
-
-
-    @ExceptionHandler({MethodArgumentTypeMismatchException.class,})
-    public ResponseEntity<Object> handleMethodArgumentTypeMismatchException(MethodArgumentTypeMismatchException ex) {
-        ApiErrorResponse response = new ApiErrorResponse();
-        response.setDescription("Incorrect argument");
-        response.setCode(String.valueOf(HttpStatus.BAD_REQUEST.value()));
-        response.setExceptionName(ex.getClass().getName());
-        response.exceptionMessage(ex.getMessage());
-        response.setStacktrace(ex.getStackTrace());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
 
     @ExceptionHandler({IncorrectRequestParameterException.class,})
@@ -55,21 +41,4 @@ public class RestExceptionHandler extends
         response.exceptionMessage(ex.getMessage());
         return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
-
-
-    @ExceptionHandler({ResourceNotFoundException.class,})
-    public ResponseEntity<Object> handleResourceNotFoundException(ResourceNotFoundException ex) {
-        ApiErrorResponse response = new ApiErrorResponse();
-        response.setDescription("Not found");
-        response.setCode(String.valueOf(HttpStatus.NOT_FOUND.value()));
-        response.setExceptionName(ex.getClass().getName());
-        response.exceptionMessage(ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-
-
-
-
-
 }
