+++ project/scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/controller/RestResponseEntityExceptionHandler.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.bot.controller.exceptionhandler;
+package ru.tinkoff.edu.java.scrapper.controller;
 
 
 import org.openapitools.model.ApiErrorResponse;
@@ -10,9 +10,10 @@ import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import ru.tinkoff.edu.java.bot.exception.IncorrectRequestParameterException;
-
+import ru.tinkoff.edu.java.scrapper.exception.IncorrectRequestParameterException;
+import ru.tinkoff.edu.java.scrapper.exception.ResourceNotFoundException;
 
 @RestControllerAdvice
 public class RestResponseEntityExceptionHandler extends
@@ -24,12 +25,24 @@ public class RestResponseEntityExceptionHandler extends
                                                                   HttpStatusCode status,
                                                                   WebRequest request) {
         ApiErrorResponse response = new ApiErrorResponse();
-        response.setDescription("Incorrect JSON");
-        response.setCode(String.valueOf(HttpStatus.BAD_REQUEST));
+        response.setDescription("Incorrect request body");
+        response.setCode(String.valueOf(HttpStatus.BAD_REQUEST.value()));
         response.setExceptionName(ex.getClass().getName());
         response.exceptionMessage(ex.getMessage());
         response.setStacktrace(ex.getStackTrace());
         return new ResponseEntity<>(response, status);
+    }
+
+
+    @ExceptionHandler({MethodArgumentTypeMismatchException.class,})
+    public ResponseEntity<Object> handleMethodArgumentTypeMismatchException(MethodArgumentTypeMismatchException ex) {
+        ApiErrorResponse response = new ApiErrorResponse();
+        response.setDescription("Incorrect argument");
+        response.setCode(String.valueOf(HttpStatus.BAD_REQUEST.value()));
+        response.setExceptionName(ex.getClass().getName());
+        response.exceptionMessage(ex.getMessage());
+        response.setStacktrace(ex.getStackTrace());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
 
     @ExceptionHandler({IncorrectRequestParameterException.class,})
@@ -41,4 +54,21 @@ public class RestResponseEntityExceptionHandler extends
         response.exceptionMessage(ex.getMessage());
         return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
+
+
+    @ExceptionHandler({ResourceNotFoundException.class,})
+    public ResponseEntity<Object> handleResourceNotFoundException(ResourceNotFoundException ex) {
+        ApiErrorResponse response = new ApiErrorResponse();
+        response.setDescription("Not found");
+        response.setCode(String.valueOf(HttpStatus.NOT_FOUND.value()));
+        response.setExceptionName(ex.getClass().getName());
+        response.exceptionMessage(ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+
+
+
+
+
 }
